requirements.txt
argh==0.30.4
GitPython==3.1.40